// src/types/serverdetail.ts
export interface ServerDetail {
  servername: string;
  serverIP: string;
  serverPort: string;
  serverStartupTime: string;
  serverWarmup: string;
  serverRoundLength: string;
  maxPlayers: string;
  status: string;
  map: string;
  mode: string;
  layer: string;
  timeStarted: string;
  players: string;
  team1: string;
  team2: string;
  tickets1: string;
  tickets2: string;
  rconUsers: string;
}